% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PioneerRWEvsRCT.R
\docType{package}
\name{PioneerRWEvsRCT-package}
\alias{PioneerRWEvsRCT}
\alias{PioneerRWEvsRCT-package}
\title{PioneerRWEvsRCT: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
